add system of lang

+++ app/src/main/java/com/example/financial_advice/app/core/model/ModelData.kt
@@ -4,7 +4,7 @@ import android.text.format.DateFormat
 import androidx.compose.ui.graphics.Color
 
 data class ModelData  (
-    var money : Int,
+    var money : Float,
     var category: String,
     var description : String? = null,
     var color : Color? = null,
